:recycle: Rename SendToken component and add docs
pass from index.js and styles.js to SendToken.js and SendToken.styles.js to improve readability
